Move compiler into its own module
