+++ extension/background/language/languageModel.test.js
@@ -1,43 +1,40 @@
 /* globals test, expect */
 
 const lm = require("./languageModel.js");
+const { compile } = require("./compiler.js");
 
 test("compiler", () => {
   expect(
-    lm
-      .compile(
-        "(bring me | take me | go | navigate | show me | open) (to | find |) (page |) [query]"
-      )
-      .toString()
+    compile(
+      "(bring me | take me | go | navigate | show me | open) (to | find |) (page |) [query]"
+    ).toString()
   ).toBe(
     'MatchPhrase("(bring me | take me | go | navigate | show me | open) (to | find | ) (page | ) [query:+]")'
   );
 
-  expect(lm.compile("clear query (database | cache)").toString()).toBe(
+  expect(compile("clear query (database | cache)").toString()).toBe(
     'MatchPhrase("clear query (database | cache)")'
   );
 
   expect(
-    lm.compile("google images (of | for |) [query] [service=images]").toString()
+    compile("google images (of | for |) [query] [service=images]").toString()
   ).toBe(
     'MatchPhrase("google images (of | for | ) [query:+]", parameters={"service":"images"})'
   );
 
   const entities = lm.convertEntities({ lang: ["Spanish", "English"] });
   expect(
-    lm
-      .compile(
-        "translate (this |) (page | tab | article | site |) to [language:lang] (for me |)",
-        entities
-      )
-      .toString()
+    compile(
+      "translate (this |) (page | tab | article | site |) to [language:lang] (for me |)",
+      entities
+    ).toString()
   ).toBe(
     'MatchPhrase("translate (this | ) (page | tab | article | site | ) to [language:(Spanish | English)] (for me | )")'
   );
 });
 
 test("basic matches", () => {
-  const phrase = lm.compile("this [query] test");
+  const phrase = compile("this [query] test");
   const results = lm.match("this is test", phrase);
   expect(results.toString()).toBe(
     'MatchResult("this is test^^", slots: {query: "is"}, capturedWords: 2)'
@@ -53,7 +50,7 @@ test("basic matches", () => {
 });
 
 test("alternative matches", () => {
-  const phrase = lm.compile("(hi | hello) world");
+  const phrase = compile("(hi | hello) world");
 
   expect(lm.match("hello world", phrase).toString()).toBe(
     'MatchResult("hello world^^", capturedWords: 2)'
@@ -69,7 +66,7 @@ test("alternative matches", () => {
 });
 
 test("Stopwords", () => {
-  const phrase = lm.compile("(launch | open) (new |) (tab | page)");
+  const phrase = compile("(launch | open) (new |) (tab | page)");
 
   expect(lm.match("launch new tab", phrase).toString()).toBe(
     'MatchResult("launch new tab^^", capturedWords: 3)'
@@ -85,7 +82,7 @@ test("Stopwords", () => {
 });
 
 test("Aliases", () => {
-  const phrase = lm.compile("(launch | open) (new |) (tab | page)");
+  const phrase = compile("(launch | open) (new |) (tab | page)");
 
   expect(lm.match("open new app", phrase).toString()).toBe(
     'MatchResult("open new app^^", aliasedWords: 1, capturedWords: 3)'
@@ -93,7 +90,7 @@ test("Aliases", () => {
 });
 
 test("Multiword aliases", () => {
-  const phrase = lm.compile("scroll upward");
+  const phrase = compile("scroll upward");
 
   expect(lm.match("scroll upward", phrase).toString()).toBe(
     'MatchResult("scroll upward^^", capturedWords: 2)'
@@ -104,7 +101,7 @@ test("Multiword aliases", () => {
 });
 
 test("Equations", () => {
-  const phrase = lm.compile("calculate [equation]");
+  const phrase = compile("calculate [equation]");
 
   expect(lm.match("calculate 2 + 3", phrase).toString()).toBe(
     'MatchResult("calculate 2 + 3^^", slots: {equation: "2 + 3"}, capturedWords: 1)'
@@ -112,8 +109,8 @@ test("Equations", () => {
 });
 
 test("Prioritizing matches", () => {
-  const fallback = lm.compile("[query]");
-  const search = lm.compile("search (for |) [query]");
+  const fallback = compile("[query]");
+  const search = compile("search (for |) [query]");
   const matchSet = new lm.MatchSet([
     ["fallback", fallback],
     ["search", search],
